### 2019-01-28 #1215 @danielskatz
Hi @edifice1989 - I note that your PDF does not have any references that are built.  It looks like you manually added references in your .md file, rather than using the `@` notation as in the [sample paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography).  Please fix this, and once you have done so, enter a new comment here with ```@whedon generate pdf``` to restart the PDF processing.
