### 2019-01-28 #1215 @arfon
> This might be an instance when replying to an email doesn't work, and you actually need to type the command in as a new comment in GitHub.

> (@arfon - do you know if this is correct?)

I think that's right. You need to actually type the command here.
