### 2019-01-29 #1215 @pjotrp
Nice contribution. I'll look for an editor.
