### 2019-02-15 #1255 @whedon
[REVIEW]: Compare Expression Profiles for Pre-defined Gene Groups with C-REx
**Submitting author:** @edifice1989 (<a href="http://orcid.org/0000-0002-8164-2480">Mingze He</a>)
**Repository:** <a href="https://github.com/edifice1989/rocker-c-rex" target ="_blank">https://github.com/edifice1989/rocker-c-rex</a>
**Version:** v1.0
