### 2019-02-15 #1255 @mgymrek
quoting from @byandell: "It is a nice application, but I think these methods are fairly standard. It allows comparison of two pre-determined groups, but often this is the start of much more intense investigation, and I don't see how to use the Shiny app beyond doing t-tests. The Shiny app appears quite polished in many ways, although I got an error with Wilcoxon test for between sample comparison, and it took awhile to initially load the data on the server."
