### 2019-02-27 #1255 @edifice1989
Hi Melissa,

    I just found a typo in the paper.md. I made changes to that. Is that
enough to fix this typo to be reflected in the final version? What else
should I do?

   Thanks!!

Best,
Ming

On Fri, Feb 15, 2019 at 2:26 PM Melissa Gymrek <notifications@github.com>
wrote:

> quoting from @byandell <https://github.com/byandell>: "It is a nice
> application, but I think these methods are fairly standard. It allows
> comparison of two pre-determined groups, but often this is the start of
> much more intense investigation, and I don't see how to use the Shiny app
> beyond doing t-tests. The Shiny app appears quite polished in many ways,
> although I got an error with Wilcoxon test for between sample comparison,
> and it took awhile to initially load the data on the server."
>
