### 2019-03-19 #1255 @edifice1989
@byandell thank you very much!
@mgymrek <https://github.com/mgymrek> please kindly advise what are next
steps. thanks!!

Best,
Ming

On Mon, Mar 18, 2019 at 12:06 PM Brian S. Yandell <notifications@github.com>
wrote:

> I filled out the checklist. Sorry, did not realize there was more to do.
> What else?
> Brian
>
> Brian S. Yandell brian.yandell@wisc.edu<mailto:brian.yandell@wisc.edu>
> Professor of Biometry, Statistics & Horticulture
> University of Wisconsin-Madison
> http://www.stat.wisc.edu/~yandell
>
>
>
> From: Mingze He <notifications@github.com>
> Reply-To: openjournals/joss-reviews <reply@reply.github.com>
> Date: Monday, March 18, 2019 at 1:07 PM
> To: openjournals/joss-reviews <joss-reviews@noreply.github.com>
> Cc: Brian Yandell <brian.yandell@wisc.edu>, Mention <
> mention@noreply.github.com>
> Subject: Re: [openjournals/joss-reviews] [REVIEW]: Compare Expression
> Profiles for Pre-defined Gene Groups with C-REx (#1255)
>
> Hi Melissa,
>
> Is there any updates from the reviewer?
>
> I am still waiting patiently: )
>
> Thanks,
> Ming
>
> On Mon, Mar 11, 2019 at 3:22 PM Melissa Gymrek <notifications@github.com>
> wrote:
>
> > @byandell <https://github.com/byandell> would you be able to fill out
> the
> > checklist?
> >
> > @edifice1989 <https://github.com/edifice1989> I will also try to
> contact
> > @byandell <https://github.com/byandell> directly
> >
> > —
> > You are receiving this because you were mentioned.
> > Reply to this email directly, view it on GitHub
> > <
> https://github.com/openjournals/joss-reviews/issues/1255#issuecomment-471763038>,
>
> > or mute the thread
> > <
> https://github.com/notifications/unsubscribe-auth/AJp68yec6ZmyBgwzg3Ii_MhUZwY6h-Qhks5vVtcSgaJpZM4a-i5Q>
>
> > .
> >
>
>
