+++ README.md
@@ -1,7 +1,12 @@
-# option one use c-rex on our lab server
+# C-REx: A shiny web application to compare RNA expression
+
+C-REx implements a novel statistical method (He et al., 2017) that was designed to assess significance of differences in RNA expression levels among specified groups of genes. This Shiny web application called C-REx (Comparison of RNA Expression) enables researchers to readily test hypotheses about whether specific gene groups share expression profiles and whether those profiles differ from those of other groups of genes. 
+
+## Installation 
+### option one use c-rex on our lab server
 http://c-rex.dill-picl.org/
 
-# option two install rocker-c-rex
+### option two install rocker-c-rex
 **Those source code are prepared to build a docker image of shiny web application called C-REx**
 
 Installation as follow:
@@ -23,3 +28,54 @@ run docker image by command line:
 
 (Note: The default port set as 3838)
 Finally the new shiny application website is available at local: 0.0.0.0:3838/app_c_rex
+
+## Quick start 
+To use C-REx, first choose to carry out a ‘within sample’ or a ‘between sample’ comparison. Next, upload or select expression input file(s) from the examples provided (see format details in the next section). For within sample comparisons, a single file is uploaded whereas between-sample comparisons require two input files. 
+
+## Input data format
+
+Comma separated values (.csv), not zipped.
+
+3 columns: 
+1. Gene ID
+2. Gene Expression value (TPM/FPKM)
+3. Gene group identifier
+
+Example:
+AC147602.5_FG004, 188, non TF genes
+AC148152.3_FG005, 8, non TF genes
+AC148152.3_FG008, 93, non TF genes
+AC148167.6_FG001, 96, non TF genes
+AC149475.2_FG002, 17, non TF genes
+AC149475.2_FG003, 37, non TF genes
+
+### Formatting Caveats
+•	Expression values can be an average from many biological replicates
+•	Do not include commas inside gene group names, a bad example would be 'Human,embryo genes'. Instead, this title should be something like this: ‘Human embryo genes’.
+•	Each annotation requires a single line. If there are, e.g., two annotations for the same gene, the same gene gets two lines, like this:
+o	AC149818.2_FG001, 188, non TF genes
+o	AC149818.2_FG001, 188, housekeeping genes
+•	If you are comparing the same group of genes under two conditions, genes with TPM or FPKM values smaller than 1 in both conditions should be filtered out. We consider such data unreliable because it is hard to tell whether the corresponding reads were from signal or noise.
+•	Use the exact label 'housekeeping genes' in the third column to annotate housekeeping genes. Some bad examples would be Housekeeping Genes, housekeeping, HOUSEKEEPING, etc.
+
+## Example usage cases
+3.1: Assessing variability between replicates
+Gene expression values are often inconsistent between biological or technical replicates (Conesa et al., 2016). C-REx can be used to carry out Student’s t-test between biological replicates to determine whether sample-based variability is so great that downstream analyses are not appropriate. 
+
+Materials and Methods
[…]
