+++ README.md
@@ -41,12 +41,18 @@ Comma separated values (.csv), not zipped.
 2. Gene Expression value (TPM/FPKM)
 3. Gene group identifier
 
-Example:
+* Example:
+
 AC147602.5_FG004, 188, non TF genes
+
 AC148152.3_FG005, 8, non TF genes
+
 AC148152.3_FG008, 93, non TF genes
+
 AC148167.6_FG001, 96, non TF genes
+
 AC149475.2_FG002, 17, non TF genes
+
 AC149475.2_FG003, 37, non TF genes
 
 ### Formatting Caveats
@@ -65,14 +71,37 @@ o	AC149818.2_FG001, 188, housekeeping genes
 •	Use the exact label 'housekeeping genes' in the third column to annotate housekeeping genes. Some bad examples would be Housekeeping Genes, housekeeping, HOUSEKEEPING, etc.
 
 ## Example usage cases
-3.1: Assessing variability between replicates
+### Assessing variability between replicates
+
 Gene expression values are often inconsistent between biological or technical replicates (Conesa et al., 2016). C-REx can be used to carry out Student’s t-test between biological replicates to determine whether sample-based variability is so great that downstream analyses are not appropriate. 
 
-Materials and Methods
+#### Materials and Methods
 RNA-seq data were collected from maize under control (non-stress) conditions (Makarevitch et al., 2015). To illustrate our method, a group of genes were selected according to annotation to GO:0006950 (response to stress) by Gramene version 37 (Tello-Ruiz et al., 2016) or maize-GAMER (Wimalanathan et al., 2018). Maize housekeeping genes were designated by Lin et al., 2014). This results in 4 files, “Gramene-non-stress-biological-replicate-1.csv”, “Gramene-non-stress-biological-replicate-2.csv”, “maize-GAMER-non-stress-biological-replicate-1.csv”, and “maize-GAMER-non-stress-biological-replicate-2.csv” (available online under the C-REx “How to” tab). Each file has 3 columns: gene ID, expression value (FPKM/TPM), and gene group name (GO:0006950 used to annotate gene group in this example). Files marked by the same annotation method are compared between biological replicates.  For instance, “Gramene-non-stress-biological-replicate-1.csv”, and “Gramene-non-stress-biological-replicate-2.csv” were uploaded to C-REx as a “Between sample comparison” session and were processed by the automatic computational pipeline (for mathematics details please refer to He et al., 2018). By choosing GO:0006950 (genes annotated by Gramene as GO:0006950) under “Choose Gene Groups” on control panel and clicking on the Student’s t-test tab on result panel, the Student’s t-test p-value is returned for the Gramene dataset. The same process applies for the cognate maize-GAMER analysis.  
 
-Results
+#### Results
 Student t-test p-values indicate that there are no significant expression differences between biological replicates under non-stressed growing conditions. This is true using the group of genes tagged by this GO term for both Gramene and maize-GAMER (Table S1). This outcome indicates that the gene groups specified in both GO annotations datasets behave consistently between replicates and findings of gene expression differences by C-REx are not likely due to sampling effects (i.e., results are not artefactual). 
+
+### Influence of gene grouping methods on observed expression differences
+
+Not all GO datasets assign the same group of genes to a given GO term. Here, we show how differences between GO annotations can influence the outcomes of gene expression analyses for both GO enrichment and for analysis using C-REx.
+
+#### Materials and Methods
+
[…]
