+++ README.md
@@ -82,7 +82,8 @@ RNA-seq data were collected from maize under control (non-stress) conditions (Ma
 Student t-test p-values indicate that there are no significant expression differences between biological replicates under non-stressed growing conditions. This is true using the group of genes tagged by this GO term for both Gramene and maize-GAMER (Table S1). This outcome indicates that the gene groups specified in both GO annotations datasets behave consistently between replicates and findings of gene expression differences by C-REx are not likely due to sampling effects (i.e., results are not artefactual). 
 
 **Table S1. t-test p-values between biological replicates.**
-|   |     Gramene     |  maize-GAMER |
+
+|   |  Gramene  |  maize-GAMER |
 |----------|:-------------:|------:|
 | p-value |  0.3497 | 0.5723 |
 
