+++ paper.bib
@@ -22,4 +22,184 @@ title = {{Transposable Elements Contribute to Activation of Maize Genes in Respo
 volume = {11},
 year = {2015}
 }
+@article{Robinson2010,
+abstract = {The fine detail provided by sequencing-based transcriptome surveys suggests that RNA-seq is likely to become the platform of choice for interrogating steady state RNA. In order to discover biologically important changes in expression, we show that normalization continues to be an essential step in the analysis. We outline a simple and effective method for performing normalization and show dramatically improved results for inferring differential expression in simulated and publicly available data sets.},
+archivePrefix = {arXiv},
+arxivId = {PMC2864565},
+author = {Robinson, Mark D and Oshlack, Alicia},
+doi = {10.1186/gb-2010-11-3-r25},
+eprint = {PMC2864565},
+isbn = {1465-6914 (Electronic)$\backslash$r1465-6906 (Linking)},
+issn = {1474-760X},
+journal = {Genome Biology},
+pages = {R25},
+pmid = {20196867},
+title = {{A scaling normalization method for differential expression analysis of RNA-seq data}},
+url = {http://dx.doi.org/10.1186/gb-2010-11-3-r25{\%}5Cnhttps://genomebiology.biomedcentral.com/articles/10.1186/gb-2010-11-3-r25{\%}5Cnhttps://genomebiology.biomedcentral.com/track/pdf/10.1186/gb-2010-11-3-r25?site=genomebiology.biomedcentral.com},
+volume = {11},
+year = {2010}
+}
+@article{Li2011,
+abstract = {BACKGROUND: RNA-Seq is revolutionizing the way transcript abundances are measured. A key challenge in transcript quantification from RNA-Seq data is the handling of reads that map to multiple genes or isoforms. This issue is particularly important for quantification with de novo transcriptome assemblies in the absence of sequenced genomes, as it is difficult to determine which transcripts are isoforms of the same gene. A second significant issue is the design of RNA-Seq experiments, in terms of the number of reads, read length, and whether reads come from one or both ends of cDNA fragments.: We present RSEM, an user-friendly software package for quantifying gene and isoform abundances from single-end or paired-end RNA-Seq data. RSEM outputs abundance estimates, 95{\%} credibility intervals, and visualization files and can also simulate RNA-Seq data. In contrast to other existing tools, the software does not require a reference genome. Thus, in combination with a de novo transcriptome assembler, RSEM enables accurate transcript quantification for species without sequenced genomes. On simulated and real data sets, RSEM has superior or comparable performance to quantification methods that rely on a reference genome. Taking advantage of RSEM's ability to effectively use ambiguously-mapping reads, we show that accurate gene-level abundance estimates are best obtained with large numbers of short single-end reads. On the other hand, estimates of the relative frequencies of isoforms within single genes may be improved through the use of paired-end reads, depending on the number of possible splice forms for each gene.$\backslash$n$\backslash$nCONCLUSIONS: RSEM is an accurate and user-friendly software tool for quantifying transcript abundances from RNA-Seq data. As it does not rely on the existence of a reference genome, it is particularly useful for quantification with de novo transcriptome assemblies. In addition, RSEM has enabled valuable guidance for cost-efficient design of quantification experiments with RNA-Seq, which is currently relatively expensive.},
+archivePrefix = {arXiv},
+arxivId = {NIHMS150003},
+author = {Li, Bo and Dewey, Colin N},
+doi = {10.1186/1471-2105-12-323},
+eprint = {NIHMS150003},
+isbn = {1471-2105 (Electronic)$\backslash$r1471-2105 (Linking)},
+issn = {1471-2105},
+journal = {BMC bioinformatics},
[…]
