+++ paper.md
@@ -39,11 +39,11 @@ bibliography: paper.bib
 Most gene expression analysis methods discover groups of genes that are co-expressed, rather than testing whether a specified gene group behaves in a concerted manner. We implemented a novel statistical method designed to assess significance of differences in RNA expression levels among specified groups of genes. Our Shiny web application C-REx (Comparison of RNA Expression) enables researchers to readily test hypotheses about whether specific gene groups share expression profiles and whether those profiles differ from those of other groups of genes. We implemented data transformation, a normality visualizer, and both parametric and non-parametric tests for determining whether gene groups are functioning in concert or in contrast both within and between conditions. Here, we demonstrate that the C-REx application recovers well-known biological phenomena (e.g., response to heat stress).
 
 # Statement of need
-RNA-seq-based gene expression levels can be variable to the extreme [@Conesa2016]. Many sophisticated methods have been devel-oped and implemented to reduce noise in datasets [@Ding2015][@Stegle2012] [@Leek2012] and to compare and define sets of differentially expressed genes (DEGs; [@Goeman2005] [@Goeman2007][@Robinson2010][@Li2011][@Trapnell2012][@Anders2010][@Patro2017]. For the most common types of expression analyses, DEG sets are iden-tified as those genes for which log2 (treatment/control) differences are > 1. Next, the identified set of DEGs is analyzed to find shared func-tional characteristics, often via gene ontology (GO) enrichment [@Thomas2003]. This results in discovery of genes that share expression profiles alongside shared functional annotations for that gene group. While this method helps to form gene groups and to figure out what functional characteristics the genes identified have in com-mon, it does not enable specific hypothesis testing. For example, if a group of twelve genes are all involved in a particular biochemical pathway, a researcher cannot use enrichment to determine whether that gene group’s expression changes are unique from other genes or gene groups. To enable this kind of assessment, we developed a method that determines whether specified groups of genes are similar (or dif-ferent) in their expression patterns [@He2018]. To do this, we changed the dimension of comparison and treated each gene as a vari-able and compare groups rather than relying on defining DEGs individ-ually. Our method uses log-transformed gene expression values, which are nearly Gaussian in distribution. If a normality requirement is met, Student’s t-test can be applied to assess the significance of differences among groups of genes between samples or treatments. If not, Wilcox-on signed-rank test can be applied. Here, we describe C-REx, an appli-cation that implements this method
[…]
