### 2019-03-31 #1255 @edifice1989
On Mon, Mar 25, 2019 at 8:57 AM Melissa Gymrek <notifications@github.com>
wrote:

> Thanks @byandell <https://github.com/byandell> for the review.
>
> @edifice1989 <https://github.com/edifice1989> I just have minor comments
> that should be addressed:
>
>    - The README.md file should be more informative, with at the very
>    least a description of what the software does, install instructions (which
>    are there already), and basic usage.
>    - For the paper, we require a statement of need. Could you add a brief
>    description of how this fulfills a need not met by existing tools, and add
>    some references to those?
>
