### 2019-03-31 #1255 @edifice1989
@mgymrek <https://github.com/mgymrek> thanks for your kind and helpful
comments!

I have updated README.md to include table of contents, more information
about the purpose of C-REx and 3 more example usages

Also, I added a section of statement of need  in paper.md with references
(updated paper.bib) to explain how C-REx is different than other
approaches, especially where enrichment analysis fails instead C-REx could
be helpful

Please let me know if there is anything else need to be addressed!

Thank you very much!!

Ming

On Mon, Mar 25, 2019 at 8:57 AM Melissa Gymrek <notifications@github.com>
wrote:

> Thanks @byandell <https://github.com/byandell> for the review.
>
> @edifice1989 <https://github.com/edifice1989> I just have minor comments
> that should be addressed:
>
>    - The README.md file should be more informative, with at the very
>    least a description of what the software does, install instructions (which
>    are there already), and basic usage.
>    - For the paper, we require a statement of need. Could you add a brief
>    description of how this fulfills a need not met by existing tools, and add
>    some references to those?
>
