### 2019-04-05 #1255 @edifice1989
Hi @mgymrek <https://github.com/mgymrek>,

   Could you please take a look of my updated README.md and paper.me with
statement of need? I am looking forward for your feedback!

   Thank you very much!

Best,
Ming
