### 2019-04-13 #1255 @byandell
Who is “you” in the message? Brian

Brian S. Yandell
Professor, Statistics & Horticulture
Director, Biometry Program
University of Wisconsin—Madison
http://www.stat.wisc.edu/~Yandell
