### 2019-04-13 #1255 @labarba
From the review thread, it looks like @mgymrek requested some changes 19 days ago: https://github.com/openjournals/joss-reviews/issues/1255#issuecomment-476261001

The author reports having addressed these.
