### 2019-04-25 #1255 @edifice1989
Hi @labarba <http://lorenabarba.com/>,

   Thanks for your inquiry! Have you heard anything back from @mgymrek
<https://github.com/mgymrek> ?

   It is almost a month since I addressed her comments.

  Thank you very much!!

On Sat, Apr 13, 2019 at 10:40 AM Lorena A. Barba <notifications@github.com>
wrote:

> From the review thread, it looks like @mgymrek
> <https://github.com/mgymrek> requested some changes 19 days ago: #1255
> (comment)
> <https://github.com/openjournals/joss-reviews/issues/1255#issuecomment-476261001>
>
> The author reports having addressed these.
>
