### 2019-04-26 #1255 @edifice1989
Here is the DOI from Zenodo: 10.5281/zenodo.2651751

Thanks!!

On Thu, Apr 25, 2019 at 1:16 PM Melissa Gymrek <notifications@github.com>
wrote:

> I apologize for the delay. Thank you for addressing the comments, this
> looks good. Can you now make a Zenodo archive and report the DOI in this
> thread?
>
