### 2019-04-29 #1255 @edifice1989
Hi @mgymrek <https://github.com/mgymrek> Does the Zenodo archive and DOI
look good? Is there anything else I need to do? thank you very much!
