### 2019-05-01 #1255 @kyleniemeyer
Hi @edifice1989, I submitted a minor PR for the paper that fixed some of the references: https://github.com/edifice1989/rocker-c-rex/pull/2

Please merge that, and we can move forward with the final acceptance.
