### 2019-05-01 #1255 @edifice1989
Hi Kyle,

     I have merged your request.

    Thank you very much for your help!!

Best,
Ming

On Tue, Apr 30, 2019 at 5:29 PM Kyle Niemeyer <notifications@github.com>
wrote:

> Hi @edifice1989 <https://github.com/edifice1989>, I submitted a minor PR
> for the paper that fixed some of the references:
> edifice1989/rocker-c-rex#2
> <https://github.com/edifice1989/rocker-c-rex/pull/2>
>
> Please merge that, and we can move forward with the final acceptance.
>
