### 2019-05-01 #1255 @kyleniemeyer
@edifice1989 congrats on your paper's acceptance in JOSS! Thanks to @byandell for reviewing and @mgymrek for editing.
